Track results of a solver run in the context
This allows for multiple ‘results’ to be generated from a single run. For
example: the best solution, as well as the ranks matrix.

It also allows any object with access to the context to publish a new 
best solution.  This may be useful to listeners that collaborate with
external sources, e.g. the AgentListener that may pull a new best solution
from the API.