Periodically preempt constraint checker
When starting from a random initial solution, it may be difficult to
find any solution that satisfies the constraints checker. At high temps
we will generate several potential solutions trying to satisfy the
constraints but will eventually choose a potential solution to score
even if it doesn’t satisfy the constraints just so the algorithm can
make progress. At low temps, we will spin until we find a solution that
satisfies the constraints.